requirements.txt
# === FastAPI 应用框架相关 ===
fastapi>=0.121.0,<0.122.0          # 主框架，提供高性能异步 API 能力
uvicorn>=0.38.0,<0.39.0            # ASGI 服务器，用于运行 FastAPI 应用
python-multipart>=0.0.20,<0.1.0    # 处理表单和文件上传（支持 multipart/form-data）

# === 图像与向量计算相关 ===
pillow>=12.0.0,<12.1.0             # 图像加载、缩放、格式转换等操作
faiss-cpu>=1.12.0,<1.13.0          # Facebook 出品的高性能相似度搜索库（CPU版）

# === 深度学习模型支持 ===
torch>=2.9.0,<2.10.0               # PyTorch 深度学习框架，用于向量化模型推理
torchvision>=0.24.0,<0.25.0        # PyTorch 的视觉库，支持图像转换、模型加载等
clip-anytorch>=2.6.0,<2.7.0        # OpenAI CLIP 模型的通用封装（兼容任何 PyTorch 版本）

# === 说明 ===
# 上述依赖用于实现一个基于 FastAPI 的图片相似度搜索系统
# 包括图片上传、特征提取（CLIP）、向量索引（FAISS）、搜索与批量管理等功能
# 如果你有 GPU：pip install faiss-gpu
